recycler view of all the ingredients and increase/decrease functionality

+++ app/src/main/java/com/example/greenplate/models/Ingredient.java
@@ -2,13 +2,13 @@ package com.example.greenplate.models;
 
 public class Ingredient {
     private String name;
-    private double quantity;
+    private int quantity;
     private int caloriesPerServing;
     private String expirationDate; // Optional, can be null if not provided
     public Ingredient() {
         //
     }
-    public Ingredient(String name, double quantity, int caloriesPerServing, String expirationDate) {
+    public Ingredient(String name, int quantity, int caloriesPerServing, String expirationDate) {
         this.name = name;
         this.quantity = quantity;
         this.caloriesPerServing = caloriesPerServing;
@@ -27,7 +27,7 @@ public class Ingredient {
         return quantity;
     }
 
-    public void setQuantity(double quantity) {
+    public void setQuantity(int quantity) {
         this.quantity = quantity;
     }
